Quick fix on defaults also including non-instance properties
The defaults override will also set as new property any new
Array.prototype override because of the “of” keyword in coffeescript.

E.g. if you define Array.prototype.sample = function() {}, sample will
then be present in the default overrides.

The way to fix is to use own, to make sure only the instance properties
of defaults is applied.